Added DiscountsUpdater class for changing weekly discounts
Added Unit tests for DiscountsUpdater, test failed due to methods not implemented.

Added methods to UpdateDiscounts

Test Passed.

Next steps:

Refactor ProductInfo to remove Discounts responsiblity

Reactor Checkout to use DiscountsInfo for discounts calculation

+++ checkout-kata.Entity/Discounts.cs
@@ -10,9 +10,9 @@ namespace checkout_kata.Entity
         
         public int ItemDiscountQuantity { get; set; }
 
-        public int ItemDiscountPrice { get; set; }
+        public double ItemDiscountPrice { get; set; }
 
-        public Discounts(string sku, int itemDiscountQuantity, int itemDiscountPrice )
+        public Discounts(string sku, int itemDiscountQuantity, double itemDiscountPrice )
         {
             this.Sku = sku;
             this.ItemDiscountQuantity = itemDiscountQuantity;
